fix: preserve float precision for fps in all code paths
Remove int() casts and fix type annotations that truncated non-integer
framerates (e.g. 29.97, 23.976 fps) to integers, causing audio/video
sync drift.

Fixes #201

+++ src/python/pose_format/numpy/pose_body.py
@@ -298,13 +298,13 @@ class NumPyPoseBody(PoseBody):
 
         return NumPyPoseBody(self.fps, new_data, confidence)
 
-    def interpolate(self, new_fps: int = None, kind='cubic'):
+    def interpolate(self, new_fps: float = None, kind='cubic'):
         """
         Interpolates the pose data to match a new frame rate.
 
         Parameters
         ----------
-        new_fps : int, optional
+        new_fps : float, optional
             The desired frame rate for interpolation.
         kind : str, optional
             The type of interpolation. Options include: "linear", "quadratic", and "cubic".
